requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langchain-google-genai==0.0.5
google-generativeai==0.3.2
pydantic==2.5.2
chromadb==0.4.22
python-dotenv==1.0.0
PyPDF2==3.0.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
streamlit==1.31.0 